fix: resolve warnings and add metadata roundtrip tests
- Gate `use whereat::at` behind cfg(feature = "encode") to fix unused
  import warning when encode feature is disabled
- Remove redundant `as u8` casts on u8 values in detect.rs
- Replace unreachable assert (u8 <= 255) with no-op in convert.rs test
- Add 15 new metadata roundtrip tests covering:
  - EXIF embedding and extraction (AVIF 4-byte offset prefix format)
  - XMP roundtrip
  - CICP color primaries and transfer characteristics
  - Content light level (HDR)
  - Mastering display colour volume (HDR)
  - Rotation (irot box, raw codes 1/3 → 90°/270°)
  - Mirror (imir box, vertical/horizontal)
  - Combined metadata roundtrip
  - decode_full() metadata extraction
  - Cooperative cancellation during encode

+++ src/detect.rs
@@ -258,9 +258,9 @@ pub fn probe(data: &[u8]) -> Result<AvifProbe, ProbeError> {
                 if colr_start + 4 <= box_end {
                     let colr_type = &data[colr_start..colr_start + 4];
                     if colr_type == b"nclx" && colr_start + 11 <= box_end {
-                        color_primaries = Some(data[colr_start + 4] as u8);
-                        transfer_characteristics = Some(data[colr_start + 6] as u8);
-                        matrix_coefficients = Some(data[colr_start + 8] as u8);
+                        color_primaries = Some(data[colr_start + 4]);
+                        transfer_characteristics = Some(data[colr_start + 6]);
+                        matrix_coefficients = Some(data[colr_start + 8]);
                         full_range = Some(data[colr_start + 10] & 0x80 != 0);
                     } else if colr_type == b"prof" || colr_type == b"rICC" {
                         has_icc = true;
